--- a/src/Quartz/Simpl/RAMJobStore.cs
+++ b/src/Quartz/Simpl/RAMJobStore.cs
@@ -1609,7 +1609,7 @@ namespace Quartz.Simpl
 
                 while (true)
                 {
-                    var tw = timeTriggers.FirstOrDefault();
+                    var tw = timeTriggers.Min;
                     if (tw == null)
                     {
                         break;