* Document default value of **MisfireThreshold**, and exception that is thrown.
* Avoid extra indirection in `GetFiredTriggerRecordId()` as `Convert.ToString(long, IFormatProvider)` in turn calls `long.ToString(IFormatProvider)` anyway.
* Update `GetTriggerKeysInternal(...)` to immediately initialize 'outList' as that is what we end up doing anyway.
* In `ApplyMisfire(TriggerWrapper tw)`, use `AddTicks(...)` instead of `AddMilliseconds(...)` to compute 'misfireTime' as a micro-optimization.
* Apply optimizations In AcquireNext(...):
  * Use **SortedSet\<T>.Min** to get first item instead of using our extension method.
  * No need to check return value of `timeTriggers.Remove(...)` as we just obtained the trigger from that list, and this while holding the instance-level lock.
  * Avoid interface calls where possible.
  * Only recalculate 'batchEnd' if more than one triggers is requested.
* Use indexer instead of enumerator to improve performance.

<details>
<summary>Benchmark results</summary>

|                                                                                                             Method | Branch |         Mean |     Error |    StdDev |  Gen 0 | Allocated |
|------------------------------------------------------------------------------------------------------------------- |--------|-------------:|----------:|----------:|-------:|----------:|
|                                                  TriggeredJobComplete_ConcurrentExecutionDisallowed_TriggersForJob | master |    172.55 ns |  0.767 ns |  0.718 ns | 0.0303 |     128 B |
|                                                  TriggeredJobComplete_ConcurrentExecutionDisallowed_TriggersForJob | PR     |    166.92 ns |  2.431 ns |  2.155 ns | 0.0303 |     128 B |
|                                                TriggeredJobComplete_ConcurrentExecutionDisallowed_NoTriggersForJob | master |    131.87 ns |  1.231 ns |  1.151 ns | 0.0382 |     160 B |
|                                                TriggeredJobComplete_ConcurrentExecutionDisallowed_NoTriggersForJob | PR     |    128.64 ns |  0.262 ns |  0.232 ns | 0.0382 |     160 B |
|                                        AcquireNextTriggers_MaxCountIsOneAndAtLeastOneMatchingTimerTriggerAvailable | master |    499.18 ns |  4.834 ns |  4.521 ns | 0.1925 |     808 B |
|                                        AcquireNextTriggers_MaxCountIsOneAndAtLeastOneMatchingTimerTriggerAvailable | PR     |    311.16 ns |  0.519 ns |  0.434 ns | 0.1717 |     720 B |
|                                        AcquireNextTriggers_MultipleTimeTriggersForJobThatAllowsConcurrentExecution | master |  4,693.05 ns | 45.691 ns | 42.739 ns | 1.3667 |   5,728 B |
|                                        AcquireNextTriggers_MultipleTimeTriggersForJobThatAllowsConcurrentExecution | PR     |  3,815.43 ns | 11.832 ns |  9.881 ns | 1.0833 |   4,536 B |
|                                           AcquireAndRelease_MultipleTriggersForJobThatDisallowsConcurrentExecution | master |  1,777.23 ns | 23.331 ns | 21.823 ns | 0.4700 |   1,976 B |
|                                           AcquireAndRelease_MultipleTriggersForJobThatDisallowsConcurrentExecution | PR     |  1,480.68 ns |  3.539 ns |  3.310 ns | 0.3700 |   1,560 B |
|                                                                                            AcquireAndFire_Misfires | master | 10,362.66 ns |170.108 ns |159.119 ns | 1.5833 |   6,632 B |
|                                                                                            AcquireAndFire_Misfires | PR     |  9,152.53 ns | 14.891 ns | 11.626 ns | 1.1533 |   4,832 B |
| AcquireAndRelease_OneTriggerForJobsThatDisallowConcurrentExecutionAndOneTriggerForJobThatAllowsConcurrentExecution | master |    890.32 ns |  5.639 ns |  5.275 ns | 0.3333 |   1,400 B |
| AcquireAndRelease_OneTriggerForJobsThatDisallowConcurrentExecutionAndOneTriggerForJobThatAllowsConcurrentExecution | PR     |    768.37 ns |  1.277 ns |  1.195 ns | 0.2922 |   1,224 B |
|                                   AcquireAndFireAndComplete_MultipleTriggersForJobsThatDisallowConcurrentExecution | master |  4,418.43 ns | 50.970 ns | 47.677 ns | 0.9967 |   4,176 B |
|                                   AcquireAndFireAndComplete_MultipleTriggersForJobsThatDisallowConcurrentExecution | PR     |  3,942.90 ns | 10.961 ns | 10.253 ns | 0.8367 |   3,512 B |
</details>

I split the changes in multiple commit for easier review.